+++ web/api_v1/routes/article_media.py
@@ -18,12 +18,11 @@ from web.helper.user import access_control
 def post_articles_id_media(article_id: int) -> Response:
     with conn.begin() as s:
         # Get article
-        # Raise if article doesn't exist
         article = s.query(Article).filter_by(id=article_id).first()
         if not article:
             return response(404, ApiText.HTTP_404)
 
-        # Generate sequence number for CDN_AUTO_NAMING
+        # Generate sequence number
         sequence = 1
         if config.CDN_AUTO_NAMING:
             last_media = (
@@ -37,6 +36,9 @@ def post_articles_id_media(article_id: int) -> Response:
                 sequence = int(match.group(1))
 
         for request_file in request.files.getlist("file"):
+            # Increment sequence
+            sequence += 1
+
             # Create details
             name, extension = os.path.splitext(request_file.filename)
             if config.CDN_AUTO_NAMING:
@@ -46,7 +48,7 @@ def post_articles_id_media(article_id: int) -> Response:
             extension = extension.lstrip(".").lower()
             filename = f"{name}.{extension}"
 
-            # Create path
+            # Create CDN path
             cdn_path_parts = ["article", article.slug, filename]
             if config.APP_DEBUG:
                 cdn_path_parts.insert(0, "_development")
@@ -60,20 +62,15 @@ def post_articles_id_media(article_id: int) -> Response:
             else:
                 continue
 
-            # Upload file
-            # Insert file
+            # Upload media
             cdn.upload(request_file, cdn_path)
+
+            # Insert file and article media
             file = File(path=cdn_path, type_id=type_id)
             s.add(file)
             s.flush()
-
-            # Insert article_media
             article_media = ArticleMedia(article_id=article_id, file_id=file.id)
             s.add(article_media)
-            s.flush()
-
-            # Increment sequence for CDN_AUTO_NAMING
-            sequence += 1
 
     return response()
 
@@ -85,25 +82,19 @@ def patch_articles_id_media_id(article_id: int, media_id: int) -> Response:
     order, has_order = json_get("order", int)
 
     with conn.begin() as s:
-        # Get article_media
-        # Raise if article_media doesn't exist
+        # Get article media and file
         article_media = (
             s.query(ArticleMedia).filter_by(id=media_id, article_id=article_id).first()
         )
         if not article_media:
             return response(404, ApiText.HTTP_404)
-
-        # Get file
-        # Raise if file doesn't exist
         file = s.query(File).filter_by(id=article_media.file_id).first()
         if not file:
             return response(404, ApiText.HTTP_404)
 
-        # Update order
+        # Update article media and file
         if has_order:
             article_media.order = order
-
-        # Update desc
         if has_desc:
             file.desc = desc
 
@@ -114,23 +105,20 @@ def patch_articles_id_media_id(article_id: int, media_id: int) -> Response:
 @api_v1_bp.delete("/articles/<int:article_id>/media/<int:media_id>")
 def delete_articles_id_media_id(article_id: int, media_id: int) -> Response:
     with conn.begin() as s:
-        # Get article_media
-        # Raise if article_media doesn't exist
+        # Get article media and file
         article_media = (
             s.query(ArticleMedia).filter_by(id=media_id, article_id=article_id).first()
         )
         if not article_media:
             return response(404, ApiText.HTTP_404)
-
-        # Get file
-        # Raise if file doesn't exist
         file = s.query(File).filter_by(id=article_media.file_id).first()
         if not file:
             return response(404, ApiText.HTTP_404)
 
-        # Remove file
-        # Delete file and article_media
+        # Delete file from CDN
         cdn.delete(file.path)
+
+        # Delete article media and file
         s.delete(file)
         s.delete(article_media)