+++ web/api_v1/routes/coupon.py
@@ -16,12 +16,12 @@ def post_coupons() -> Response:
     rate = abs(percentage / 100 - 1) if percentage else None
 
     with conn.begin() as s:
-        # Raise if coupon exists
+        # Check if coupon already exists
         coupon = s.query(Coupon).filter_by(code=code, is_deleted=False).first()
         if coupon:
             return response(409, ApiText.HTTP_409)
 
-        # Add coupon
+        # Insert coupon
         coupon = Coupon(code=code, rate=rate, amount=amount)
         s.add(coupon)
 
@@ -32,13 +32,10 @@ def post_coupons() -> Response:
 @api_v1_bp.delete("/coupons/<int:coupon_id>")
 def delete_coupons_id(coupon_id: int) -> Response:
     with conn.begin() as s:
-        # Get coupon
-        # Raise if coupon doesn't exist
+        # Delete coupon
         coupon = s.query(Coupon).filter_by(id=coupon_id).first()
         if not coupon:
             return response(404, ApiText.HTTP_404)
-
-        # Update is_deleted
         coupon.is_deleted = True
 
     return response()