+++ web/api_v1/routes/email.py
@@ -27,5 +27,5 @@ def post_emails_contact() -> Response:
         send_custom_1(**data)
         send_contact_customer(email=data["email"], message=data["message"])
         return response(200, _Text.CONTACT_SUCCESS)
-    else:
-        return response(403, _Text.INVALID_TEMPLATE_ID)
+
+    return response(403, _Text.INVALID_TEMPLATE_ID)